requirements.txt
azure-cognitiveservices-speech==1.36.0
dateparser==1.1.8
pathlib==1.0.1
pydub==0.25.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
regex==2023.8.8
six==1.16.0
tzlocal==5.0.1
wave==0.0.2
python-ffmpeg==2.0.12
